+++ .idea/.name
@@ -1,1 +1,1 @@
-Pragyan-main-app
+pragyan16
