+++ .idea/gradle.xml
@@ -1,14 +1,19 @@
-   <component name="GradleSettings">
-      <option name="linkedExternalProjectsSettings">
-        <GradleProjectSettings>
-         <option name="distributionType" value="DEFAULT_WRAPPED" />
-          <option name="externalProjectPath" value="$PROJECT_DIR$" />
-          <option name="modules">
-            <set>
-              <option value="$PROJECT_DIR$" />
- @@ -16,4 +14,5 @@
-        </GradleProjectSettings>
-      </option>
-    </component>
- </project>
- 
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="GradleSettings">
+    <option name="linkedExternalProjectsSettings">
+      <GradleProjectSettings>
+        <option name="distributionType" value="LOCAL" />
+        <option name="externalProjectPath" value="$PROJECT_DIR$" />
+        <option name="gradleHome" value="$APPLICATION_HOME_DIR$/gradle/gradle-2.2.1" />
+        <option name="gradleJvm" value="1.8" />
+        <option name="modules">
+          <set>
+            <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/app" />
+          </set>
+        </option>
+      </GradleProjectSettings>
+    </option>
+  </component>
+</project>
