+++ .idea/misc.xml
@@ -1,112 +1,22 @@
- <?xml version="1.0" encoding="UTF-8"?>
-  <project version="4">
-   <component name="DaemonCodeAnalyzer">
-     <disable_hints />
-   </component>
-    <component name="EntryPointsManager">
-      <entry_points version="2.0" />
-    </component>
- @@ -13,10 +16,103 @@
-      <ConfirmationsSetting value="0" id="Add" />
-      <ConfirmationsSetting value="0" id="Remove" />
-    </component>
-   <component name="ProjectRootManager" version="2" languageLevel="JDK_1_7" assert-keyword="true" jdk-15="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
-      <output url="file://$PROJECT_DIR$/build/classes" />
-    </component>
-   <component name="RunManager">
-     <configuration default="true" type="JUnit" factoryName="JUnit">
-       <module name="" />
-       <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-       <option name="ALTERNATIVE_JRE_PATH" />
-       <option name="PACKAGE_NAME" />
-       <option name="MAIN_CLASS_NAME" />
-       <option name="METHOD_NAME" />
-       <option name="TEST_OBJECT" value="class" />
-       <option name="VM_PARAMETERS" value="-ea" />
-       <option name="PARAMETERS" />
-       <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
-       <option name="ENV_VARIABLES" />
-       <option name="PASS_PARENT_ENVS" value="true" />
-       <option name="TEST_SEARCH_SCOPE">
-         <value defaultName="moduleWithDependencies" />
-       </option>
-       <envs />
-      <patterns />
-       <method />
-     </configuration>
-     <configuration default="true" type="Remote" factoryName="Remote">
-       <option name="USE_SOCKET_TRANSPORT" value="true" />
-       <option name="SERVER_MODE" value="false" />
-       <option name="SHMEM_ADDRESS" value="javadebug" />
-       <option name="HOST" value="localhost" />
-       <option name="PORT" value="5005" />
-       <method />
-     </configuration>
-     <configuration default="true" type="Application" factoryName="Application">
-       <option name="MAIN_CLASS_NAME" />
-       <option name="VM_PARAMETERS" />
-       <option name="PROGRAM_PARAMETERS" />
-       <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
-       <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-       <option name="ALTERNATIVE_JRE_PATH" />
-       <option name="ENABLE_SWING_INSPECTOR" value="false" />
-       <option name="ENV_VARIABLES" />
-       <option name="PASS_PARENT_ENVS" value="true" />
-       <module name="" />
-       <envs />
-       <method />
-     </configuration>
-     <configuration default="true" type="TestNG" factoryName="TestNG">
-       <module name="" />
-       <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-       <option name="ALTERNATIVE_JRE_PATH" />
-       <option name="SUITE_NAME" />
-       <option name="PACKAGE_NAME" />
-       <option name="MAIN_CLASS_NAME" />
-       <option name="METHOD_NAME" />
-       <option name="GROUP_NAME" />
-       <option name="TEST_OBJECT" value="CLASS" />
-       <option name="VM_PARAMETERS" value="-ea" />
-       <option name="PARAMETERS" />
-       <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
-       <option name="OUTPUT_DIRECTORY" />
-       <option name="ANNOTATION_TYPE" />
-       <option name="ENV_VARIABLES" />
-       <option name="PASS_PARENT_ENVS" value="true" />
-       <option name="TEST_SEARCH_SCOPE">
-         <value defaultName="moduleWithDependencies" />
-       </option>
-       <option name="USE_DEFAULT_REPORTERS" value="false" />
-       <option name="PROPERTIES_FILE" />
-       <envs />
-       <properties />
-       <listeners />
-       <method />
-     </configuration>
-     <list size="0" />
-     <configuration name="&lt;template&gt;" type="#org.jetbrains.idea.devkit.run.PluginConfigurationType" default="true" selected="false">
-       <option name="VM_PARAMETERS" value="-Xmx512m -Xms256m -XX:MaxPermSize=250m -ea" />
-     </configuration>
-     <configuration name="&lt;template&gt;" type="WebApp" default="true" selected="false">
-       <Host>localhost</Host>
-       <Port>5050</Port>
-     </configuration>
-   </component>
-   <component name="masterDetails">
-     <states>
-       <state key="ProjectJDKs.UI">
-         <settings>
-           <last-edited>1.8</last-edited>
-           <splitter-proportions>
-             <option name="proportions">
-               <list>
-                 <option value="0.33333334" />
-               </list>
-             </option>
-           </splitter-proportions>
-         </settings>
-       </state>
-     </states>
-    </component>
- </project>
- 
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="EntryPointsManager">
+    <entry_points version="2.0" />
+  </component>
+  <component name="ProjectLevelVcsManager" settingsEditedManually="false">
+    <OptionsSetting value="true" id="Add" />
+    <OptionsSetting value="true" id="Remove" />
+    <OptionsSetting value="true" id="Checkout" />
+    <OptionsSetting value="true" id="Update" />
+    <OptionsSetting value="true" id="Status" />
+    <OptionsSetting value="true" id="Edit" />
+    <ConfirmationsSetting value="0" id="Add" />
+    <ConfirmationsSetting value="0" id="Remove" />
+  </component>
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_7" default="true" assert-keyword="true" jdk-15="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
+    <output url="file://$PROJECT_DIR$/build/classes" />
+  </component>
+  <component name="ProjectType">
+    <option name="id" value="Android" />
+  </component>
+</project>
