+++ .idea/modules.xml
@@ -1,9 +1,9 @@
- <project version="4">
-    <component name="ProjectModuleManager">
-      <modules>
-       <module fileurl="file://$PROJECT_DIR$/Pragyan-main-app.iml" filepath="$PROJECT_DIR$/Pragyan-main-app.iml" />
-        <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      </modules>
-    </component>
- </project>
- 
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/pragyan16.iml" filepath="$PROJECT_DIR$/pragyan16.iml" />
+    </modules>
+  </component>
+</project>
