--- Pragyan-main-app.iml
@@ -1,21 +1,0 @@
- <?xml version="1.0" encoding="UTF-8"?>
- <module external.linked.project.path="$MODULE_DIR$" external.root.project.path="$MODULE_DIR$" external.system.id="GRADLE" type="JAVA_MODULE" version="4">
-    <component name="FacetManager">
-      <facet type="java-gradle" name="Java-Gradle">
-        <configuration>
-          <option name="BUILD_FOLDER_PATH" value="$MODULE_DIR$/build" />
-        </configuration>
-      </facet>
-    </component>
-   <component name="NewModuleRootManager" inherit-compiler-output="false">
-     <output url="file://$MODULE_DIR$/build/classes/main" />
-     <output-test url="file://$MODULE_DIR$/build/classes/test" />
-      <exclude-output />
-      <content url="file://$MODULE_DIR$">
-        <excludeFolder url="file://$MODULE_DIR$/.gradle" />
- @@ -16,4 +17,5 @@
-      <orderEntry type="inheritedJdk" />
-      <orderEntry type="sourceFolder" forTests="false" />
-    </component>
- </module>
-
