+++ app/src/main/AndroidManifest.xml
@@ -1,45 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.delta.pragyan16">
-
-    <uses-permission android:name="android.permission.INTERNET" />
-
+    package="com.delta.pragyan16" >
+<uses-permission android:name="android.permission.INTERNET"/>
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme" >
         <activity
             android:name=".MainActivity"
-            android:label="@string/app_name"
-            android:theme="@android:style/Theme.NoTitleBar.Fullscreen">
-
+            android:label="@string/app_name" >
         </activity>
         <activity
-            android:name=".SplashScreen"
-            android:label="@string/app_name"
-            android:theme="@android:style/Theme.NoTitleBar.Fullscreen" >
+            android:name=".EventsActivity"
+            android:label="@string/title_activity_events" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-
-        </activity>
-        <activity
-            android:name=".ClustersActivity"
-            android:label="@string/title_activity_events" />
-        <activity
-            android:name=".EventsActivity"
-            android:label="@string/title_activity_events" />
-        
-        <activity
-            android:name=".EventDetailActivity"
-            android:label="@string/title_activity_event_detail"></activity>
-        <activity
-            android:name=".SplachActivity2"
-            android:label="@string/app_name" >
-
         </activity>
     </application>
