--- app/src/main/java/com/delta/pragyan16/SplachActivity2.java
@@ -1,51 +1,0 @@
-package com.delta.pragyan16;
-
-import android.app.Activity;
-import android.graphics.Canvas;
-import android.graphics.Paint;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.SurfaceView;
-
-
-public class SplachActivity2 extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash_2);
-
-        Utilities.init_colors();
-        SurfaceView s= (SurfaceView) findViewById(R.id.fracview);
-        Paint paint=new Paint();
-        paint.setColor(Utilities.colors[0]);
-
-        Canvas canvas=new Canvas();
-        canvas.drawARGB(255,50,50,50);
-        s.draw(canvas);
-    }
-
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-}
