--- app/src/main/res/anim/pragyan_logo_bounce.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android">
-    <scale
-        android:interpolator="@android:anim/overshoot_interpolator"
-        android:fromXScale="0"
-        android:toXScale="1"
-        android:fromYScale="0"
-        android:toYScale="1.0"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        android:duration="2000" />
-</set>
