--- app/src/main/res/anim/zoom_chakra_in.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shareInterpolator="false">
-
-    <scale xmlns:android="http://schemas.android.com/apk/res/android"
-        android:interpolator="@android:anim/linear_interpolator"
-        android:duration="1000"
-        android:fillEnabled="true"
-        android:fillBefore="false"
-        android:fillAfter="true"
-        android:fromXScale="1"
-        android:fromYScale="1"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        android:toXScale="20"
-        android:toYScale="20" />
-<!--
-    <scale
-        android:interpolator="@android:anim/accelerate_decelerate_interpolator"
-        android:fromXScale="5"
-        android:toXScale="1.4"
-        android:fromYScale="5"
-        android:toYScale="0.6"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        android:fillAfter="false"
-        android:startOffset="3000"
-        android:duration="700" />
-    <set
-        android:interpolator="@android:anim/accelerate_interpolator"
-        android:startOffset="3200">
-        <scale
-            android:fromXScale="1.4"
-            android:toXScale="0.0"
-            android:fromYScale="0.6"
-            android:toYScale="0.0"
-            android:pivotX="50%"
-            android:pivotY="50%"
-            android:duration="400" />
-        <rotate
-            android:fromDegrees="0"
-            android:toDegrees="-45"
-            android:toYScale="0.0"
-            android:pivotX="50%"
-            android:pivotY="50%"
-            android:duration="400" />
-    </set>
-    -->
-</set>
