--- app/src/main/res/anim/zoom_chakra_out_left.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android"
-
-    android:shareInterpolator="false">
-
-    <set
-        android:interpolator="@android:anim/accelerate_interpolator"
-        android:fillEnabled="true"
-        android:fillBefore="false"
-        android:fillAfter="false"
-        >
-        <scale
-            android:fromXScale="20"
-            android:toXScale="0.0"
-            android:fromYScale="20"
-            android:toYScale="0.0"
-            android:pivotX="50%"
-            android:pivotY="50%"
-            android:duration="1000" />
-        <rotate
-            android:fromDegrees="0"
-            android:toDegrees="-180"
-            android:toYScale="0.0"
-            android:pivotX="50%"
-            android:pivotY="50%"
-            android:duration="1000" />
-    </set>
-</set>
