+++ app/src/main/res/layout/activity_events.xml
@@ -4,13 +4,9 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".ClustersActivity">
+    tools:context="com.delta.pragyan16.EventsActivity">
 
-    <GridView
-        android:id="@+id/clustersGrid"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:numColumns="auto_fit"
-        android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true" />
+    <TextView android:text="@string/hello_world" android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
+
 </RelativeLayout>
