--- app/src/main/res/layout/fragment_details.xml
@@ -1,51 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:layout_height="match_parent"
-    tools:context=".DetailsFragment">
-    <ScrollView
-        android:id="@+id/scrollView"
-        android:layout_weight="6"
-        android:paddingLeft="5dp"
-        android:paddingRight="5dp"
-        android:paddingBottom="3dp"
-        android:layout_width="match_parent"
-        android:layout_height="0dp">
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <TextView
-                android:id="@+id/detailText"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textAppearance="?android:attr/textAppearanceLarge"
-                android:text="Large Text"
-                android:layout_marginTop="20dp"
-                android:layout_marginBottom="25dp"
-                android:layout_alignParentTop="true"
-                android:layout_centerHorizontal="true" />
-
-            <TextView
-                android:id="@+id/descriptionDetail"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_margin="3dp"
-                android:textAppearance="?android:attr/textAppearanceMedium"
-                android:text="Medium Text"
-                android:layout_below="@+id/detailText"
-                android:layout_marginTop="75dp"/>
-
-        </RelativeLayout>
-
-    </ScrollView>
-
-    <Button
-        android:id="@+id/eventRegisterButton"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:background="#3f85f4"
-        android:textColor="#ffffff"
-        android:text="@string/register"
-        android:layout_weight="1" />
-</LinearLayout>
