+++ app/src/main/res/values/strings.xml
@@ -4,11 +4,4 @@
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
     <string name="title_activity_events">EventsActivity</string>
-    <string name="title_activity_event_details">EventDetails</string>
-    <string name="title_activity_event_detail">Event Details</string>
-    <string name="section_format">Hello World from section: %1$d</string>
-
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="register">Register</string>
 </resources>
